backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "arize-phoenix>=11.18.0",
    "arxiv>=2.2.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "llama-index>=0.13.0",
    "llama-index-callbacks-arize-phoenix>=0.6.0",
    "llama-index-vector-stores-qdrant>=0.7.1",
    "openai>=1.99.0",
    "opentelemetry-api>=1.36.0",
    "passlib>=1.7.4",
    "psycopg>=3.2.9",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pypdf2>=3.0.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]